Use generic variables for orgs and enterprises
The existing variables are too specific to the GitHub org. This commit
addresses this in the following files:

- org-repos-fragment-directive-2.graphql
- org-repos-fragment-directive.graphql

---

In addition, the following files have references to variable usage, but
was not explicitly declared anywhere:

- enterprise-saml-identities.graphql
- org-branches-and-commits-by-repository.graphql
- org-members-by-team.graphql
- org-pr-merged-info-by-repository.graphql
- repo-get-all-branches.graphql
- repos-get-last-issue-comment.graphql

I've updated the above files to just use simple strings to replace prior
to using the queries in them.